requirements.txt
Flask
pandas
numpy
mysql-connector-python
